aws bucket errors

+++ db/library/dbscript/database.php
@@ -577,8 +577,11 @@ class Database {
     $s3 = new S3( environment('awsAccessKey'), environment('awsSecretKey') );
     if (!$s3)
       trigger_error( 'Sorry, there was a problem connecting to Amazon Web Services', E_USER_ERROR );
-    if (!($s3->deleteObject(environment('awsBucket'), urlencode($aws_file))))
-      trigger_error( 'Sorry, there was a problem deleting the file from Amazon Web Services', E_USER_ERROR );
+    if ($s3->bucketExists(environment('awsBucket'))
+    && $s3->objectExists(environment('awsBucket'),urlencode($aws_file))) {
+      if (!($s3->deleteObject(environment('awsBucket'), urlencode($aws_file))))
+        trigger_error( 'Sorry, there was a problem deleting the file from Amazon Web Services', E_USER_ERROR );
+    }
   }
   
   
@@ -589,9 +592,11 @@ class Database {
     $s3 = new S3( environment('awsAccessKey'), environment('awsSecretKey') );
     if (!$s3)
       trigger_error( 'Sorry, there was a problem connecting to Amazon Web Services', E_USER_ERROR );
-    $result = $s3->putBucket( environment('awsBucket'), 'public-read' );
-    if (!$result)
-      trigger_error( 'Sorry, there was a problem creating the bucket '.environment('awsBucket').' at Amazon Web Services', E_USER_ERROR );
+    if (!$s3->bucketExists(environment('awsBucket'))) {
+      $result = $s3->putBucket( environment('awsBucket'), 'public-read' );
+      if (!$result)
+        trigger_error( 'Sorry, there was a problem creating the bucket '.environment('awsBucket').' at Amazon Web Services', E_USER_ERROR );
+    }
     if (file_exists($this->file_upload[1])) {
       if (!($s3->putObjectFile( $this->file_upload[1] , environment('awsBucket'), $file, 'public-read' )))
         trigger_error( 'Sorry, there was a problem uploading the file to Amazon Web Services', E_USER_ERROR );
